requirements.txt
python-nmap
requests
shodan
dnspython
whois
networkx
pyvis
jinja2
colorama
tqdm
validators
